+++ src/datasets/download/__init__.py
@@ -1,0 +1,9 @@
+__all__ = [
+    # "DownloadConfig",
+    "DownloadManager",
+    "DownloadMode",
+    "StreamingDownloadManager",
+]
+
+from .download_manager import DownloadManager, DownloadMode
+from .streaming_download_manager import StreamingDownloadManager
