+++ src/datasets/features/audio.py
@@ -7,9 +7,9 @@ import pyarrow as pa
 from packaging import version
 
 from .. import config
+from ..download.streaming_download_manager import xopen
 from ..table import array_cast
 from ..utils.py_utils import no_op_if_value_is_null, string_to_dict
-from ..utils.streaming_download_manager import xopen
 
 
 if TYPE_CHECKING:
