+++ src/datasets/features/image.py
@@ -7,10 +7,10 @@ import numpy as np
 import pyarrow as pa
 
 from .. import config
+from ..download.streaming_download_manager import xopen
 from ..table import array_cast
 from ..utils.file_utils import is_local_path
 from ..utils.py_utils import first_non_null_value, no_op_if_value_is_null, string_to_dict
-from ..utils.streaming_download_manager import xopen
 
 
 if TYPE_CHECKING:
