+++ src/datasets/utils/__init__.py
@@ -17,17 +17,11 @@
 """Util import."""
 
 __all__ = [
-    "DownloadConfig",
-    "DownloadManager",
-    "DownloadMode",
     "disable_progress_bar",
     "enable_progress_bar",
     "is_progress_bar_enabled",
-    "StreamingDownloadManager",
     "Version",
 ]
 
-from .download_manager import DownloadConfig, DownloadManager, DownloadMode
 from .logging import disable_progress_bar, enable_progress_bar, is_progress_bar_enabled
-from .streaming_download_manager import StreamingDownloadManager
 from .version import Version
